+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Added
 
+- concord-server: timeout options for GIT's HTTPS and SSH transports; 
 - concord-server: a policy rule for setting the maximum allowed
 `processTimeout` value.